+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.7.6] - 01-25-25
+- Added Master, InstanceOwner, FirstJoin, and FirstJoinName properties to AccessControl
+- Fixed array overflow in AccessControl if more than 100 users are in world
+
 ## [1.7.5] - 01-20-25
 - Exposed OwnerName and ComponentName in AccessEventBase
 - Fixed exception in _ReleaseImage(claimToken) in ImageDownloadManager